+++ src/MeasureMap/SessionHandlers/BasicSessionHandler.cs
@@ -19,10 +19,10 @@ namespace MeasureMap
         {
             var runnerThreads = new WorkerThreadList();
             
-            var thread = runnerThreads.StartNew(0, () =>
+            var thread = runnerThreads.StartNew(0, _ =>
             {
                 var worker = new Worker();
-                return worker.Run(task, settings);
+                return worker.Run(task, new ExecutionContext(settings));
             }, settings.GetThreadFactory());
             
             settings.Logger.Write($"Start thread {thread.Id}", LogLevel.Debug, nameof(BasicSessionHandler));