+++ src/Tests/MeasureMap.UnitTest/Threading/WorkerThreadListTests.cs
@@ -14,7 +14,7 @@ namespace MeasureMap.UnitTest.Threading
         public void WorkerThreadList_StartNew_Thread()
         {
             var list = new WorkerThreadList();
-            var thread = list.StartNew(1, () => new Result(), WorkerThread.Factory);
+            var thread = list.StartNew(1, _ => new Result(), WorkerThread.Factory);
 
             thread.Should().NotBeNull();
         }
@@ -23,7 +23,7 @@ namespace MeasureMap.UnitTest.Threading
         public void WorkerThreadList_StartNew_Task()
         {
             var list = new WorkerThreadList();
-            var thread = list.StartNew(1, () => new Result(), WorkerTask.Factory);
+            var thread = list.StartNew(1, _ => new Result(), WorkerTask.Factory);
 
             thread.Should().NotBeNull();
         }
@@ -32,7 +32,7 @@ namespace MeasureMap.UnitTest.Threading
         public void WorkerThreadList_StartNew_Added()
         {
             var list = new WorkerThreadList();
-            var thread = list.StartNew(1, () => new Result(), WorkerThread.Factory);
+            var thread = list.StartNew(1, _ => new Result(), WorkerThread.Factory);
 
             list.Single().Should().BeSameAs(thread);
         }
@@ -41,7 +41,7 @@ namespace MeasureMap.UnitTest.Threading
         public void WorkerThreadList_StartNew_Task_Added()
         {
             var list = new WorkerThreadList();
-            var thread = list.StartNew(1, () => new Result(), WorkerTask.Factory);
+            var thread = list.StartNew(1, _ => new Result(), WorkerTask.Factory);
 
             list.Single().Should().BeSameAs(thread);
         }
@@ -49,17 +49,17 @@ namespace MeasureMap.UnitTest.Threading
         [Test]
         public void WorkerThreadList_Add()
         {
-            var thread = new WorkerThread(1, () => new Result());
+            var thread = new WorkerThread(1, _ => new Result());
             var list = new WorkerThreadList();
             list.Add(thread);
-            
+
             list.Single().Should().BeSameAs(thread);
         }
 
         [Test]
         public void WorkerThreadList_Add_ListInitializer()
         {
-            var thread = new WorkerThread(1, () => new Result());
+            var thread = new WorkerThread(1, _ => new Result());
             var list = new WorkerThreadList
             {
                 thread
@@ -71,7 +71,7 @@ namespace MeasureMap.UnitTest.Threading
         [Test]
         public void WorkerThreadList_Remove()
         {
-            var thread = new WorkerThread(1, () => new Result());
+            var thread = new WorkerThread(1, _ => new Result());
             var list = new WorkerThreadList
             {
                 thread
@@ -83,7 +83,7 @@ namespace MeasureMap.UnitTest.Threading
         [Test]
         public void WorkerThreadList_WaitAll()
         {
-            var thread = new WorkerThread(1, () =>
+            var thread = new WorkerThread(1, _ =>
             {
                 System.Threading.Tasks.Task.Delay(TimeSpan.FromSeconds(1)).Wait();
                 return new Result();
@@ -102,7 +102,7 @@ namespace MeasureMap.UnitTest.Threading
         [Test]
         public void WorkerThreadList_Task_WaitAll()
         {
-            var thread = new WorkerTask(1, () =>
+            var thread = new WorkerTask(1, _ =>
             {
                 System.Threading.Tasks.Task.Delay(TimeSpan.FromSeconds(1)).Wait();
                 return new Result();
@@ -122,12 +122,12 @@ namespace MeasureMap.UnitTest.Threading
         public void WorkerThreadList_WaitAll_MultipleTreads()
         {
             var list = new WorkerThreadList();
-            list.StartNew(1, () =>
+            list.StartNew(1, _ =>
             {
                 System.Threading.Tasks.Task.Delay(TimeSpan.FromSeconds(1)).Wait();
                 return new Result();
             }, WorkerThread.Factory).Start();
-            list.StartNew(1, () =>
+            list.StartNew(1, _ =>
             {
                 System.Threading.Tasks.Task.Delay(TimeSpan.FromSeconds(1)).Wait();
                 return new Result();
@@ -142,12 +142,12 @@ namespace MeasureMap.UnitTest.Threading
         public void WorkerThreadList_WaitAll_MultipleTasks()
         {
             var list = new WorkerThreadList();
-            list.StartNew(1, () =>
+            list.StartNew(1, _ =>
             {
                 System.Threading.Tasks.Task.Delay(TimeSpan.FromSeconds(1)).Wait();
                 return new Result();
             }, WorkerTask.Factory).Start();
-            list.StartNew(1, () =>
+            list.StartNew(1, _ =>
             {
                 System.Threading.Tasks.Task.Delay(TimeSpan.FromSeconds(1)).Wait();
                 return new Result();